# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/notification/get-all-notified-emails": {
            "post": {
                "tags": [
                    "notification-controller-impl"
                ],
                "summary": "Get all notified email list",
                "operationId": "getAllNotifiedEmails",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetAllNotifiedEmailsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Got all notified email list failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Got all notified email list successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAllNotifiedEmailsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "stackTrace": {
                        "type": "string"
                    }
                }
            },
...
            "GetAllNotifiedEmailsRequest": {
                "type": "object",
                "properties": {
                    "sender": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    }
                }
            },
            "GetAllNotifiedEmailsResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "recipients": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
        }
    }
}